+++ klibs/KLCommunication.py
@@ -81,11 +81,10 @@ def collect_demographics(anonymous=False):
 		if q.active:
 			# if querying unique identifier, make sure it doesn't already exist in db
 			if q.database_field == P.unique_identifier:
-				# TODO: fix this to work with multi-user mode
-				existing = db.query("SELECT `{0}` FROM `participants`".format(q.database_field))
+				existing = [utf8(pid) for pid in db.get_unique_ids()]
 				while True:
 					value = query(q, anonymous=anonymous)
-					if utf8(value) in [utf8(val[0]) for val in existing]:
+					if utf8(value) in existing:
 						err = ("A participant with that ID already exists!\n"
 								"Please try a different identifier.")
 						fill()
@@ -119,12 +118,6 @@ def collect_demographics(anonymous=False):
 	else:
 		#  The context for this is: collect_demographics is set to false but then explicitly called later
 		db.update(demographics.table, demographics.defined)
-
-	if P.multi_session_project and not P.manual_demographics_collection:
-		try:
-			exp.init_session()
-		except:
-			pass
 
 
 def init_default_textstyles():